Revert a patch that dropped user profile on platform change.
The patch was proven not to work. Eclipse believed that the no JDT and
PDE was installed (nor anything that was installed via root dropins) and
fetched those bundles again from update sites if they were required.
This was a very unwanted behavior, because we lost control over what is
and what is not in the installation.